requirements.txt
fastapi 
uvicorn 
web3 
python-dotenv 
python-jose 
passlib[bcrypt]